+++ novae/__init__.py
@@ -11,7 +11,7 @@ from . import data
 from . import monitor
 from . import plot
 from .utils import spatial_neighbors, batch_effect_correction
-from .data import load_dataset
+from .data import load_dataset, toy_dataset, quantile_scaling, compute_histo_embeddings, compute_histo_pca
 
 log = logging.getLogger("novae")
 configure_logger(log)