+++ novae/data/__init__.py
@@ -2,6 +2,5 @@ from .convert import AnnDataTorch
 from .dataset import NovaeDataset
 from .datamodule import NovaeDatamodule
 from ._embeddings._histo import compute_histo_embeddings, compute_histo_pca
-from .load import _load_wandb_artifact, load_dataset
+from ._load import _load_wandb_artifact, load_dataset, toy_dataset
 from .preprocess import quantile_scaling
-from .toy import toy_dataset